move out components to new folder and fix tests
