+++ static/app/components/modals/sentryAppPublishRequestModal/sentryAppUtils.tsx
