+++ static/app/views/settings/organizationDeveloperSettings/sentryApplicationRow/index.tsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled';
 
 import {openModal} from 'sentry/actionCreators/modal';
 import Link from 'sentry/components/links/link';
-import {SentryAppPublishRequestModal} from 'sentry/components/modals/sentryAppPublishRequestModal';
+import {SentryAppPublishRequestModal} from 'sentry/components/modals/sentryAppPublishRequestModal/sentryAppPublishRequestModal';
 import PanelItem from 'sentry/components/panels/panelItem';
 import SentryAppIcon from 'sentry/components/sentryAppIcon';
 import {space} from 'sentry/styles/space';
